feat: :sparkles: Added an option to recover unuploaded files (Not tested !!)
Now, you can enable an option to recover unuploaded files. This can happen sometimes with a time out error. Now, the files are stored in a data file so every non-uploaded files can be uploaded next time

+++ index.js
@@ -19,7 +19,11 @@ if(googlePhotos.authToken == "" || googlePhotos.authToken == null) {
     process.exit(0);
 }
 
-init();
+if(config.application.tryToRecoverUnuploadedFiles) {
+    checkUnuploadedFiles();
+} else {
+    init();
+}
 
 async function init() {
     let files = [];
@@ -34,10 +38,33 @@ async function init() {
             });
         });
     })
-    console.info("There are " + files.length + " files");
+    console.info("There are " + files.length + " files in your OneDrive folder");
     files.forEach(file => {
         downloadOneDriveFiles(file.id, file.name);
     });
+}
+
+function checkUnuploadedFiles() {
+    console.info("Checking if files were not uploaded last time...");
+    let fileNameList = [];
+    let filesNumber = 0;
+    fs.readdirSync(__dirname + "/downloads/").forEach(file => {
+        if(!file.endsWith('recovery.json')) {
+            filesNumber++;
+            fileNameList.push(file);
+        }
+    });
+    if(filesNumber == 0) {
+        console.info("Perfect ! Last time, every files were uploaded");
+    } else {
+        console.info("Last time, " + filesNumber + " file(s) was/were not uploaded");
+        fileNameList.forEach(element => {
+            uploadToGooglePhotos(element);
+        });
+    }
+
+    init();
+    
 }
 
 /** Downloading data from OneDrive */
@@ -51,6 +78,7 @@ async function downloadOneDriveFiles(fileId, fileName) {
     fileStream.pipe(stream);
     await once(fileStream, "end");
     console.info("File " + fileName + " downloaded");
+    addFileToUploadList(fileName);
     setTimeout(() => {
         uploadToGooglePhotos(fileName, fileId);
     }, 1500);
@@ -58,20 +86,31 @@ async function downloadOneDriveFiles(fileId, fileName) {
 
 /** Uploading data to Google Photos */
 
-async function uploadToGooglePhotos(fileName, fileId) {
-    try {
-        let uploadToken = await photos.transport.upload(fileName, __dirname + "/downloads/" + fileName, config.application.maxUploadingTime * 1000);
-        let upload = await photos.mediaItems.albumBatchCreate(googlePhotos.albumId, fileName, "Automatic upload", uploadToken)
-        console.info("File " + fileName +" uploaded");
-        deleteFile(fileName, fileId);
-    } catch (error) {
-        console.info("File " + fileName + " was not uploaded (Time out, try increasing the maxUploadingTime value in the config)")
+async function uploadToGooglePhotos(fileName, fileId = NaN) {
+
+    if(getUploadList().contains(fileName)) {
+        try {
+            let uploadToken = await photos.transport.upload(fileName, __dirname + "/downloads/" + fileName, config.application.maxUploadingTime * 1000);
+            let upload = await photos.mediaItems.albumBatchCreate(googlePhotos.albumId, fileName, "Automatic upload", uploadToken)
+            console.info("File " + fileName +" uploaded");
+            deleteFile(fileName, fileId);
+            removeFileFromUploadList(fileName);
+        } catch (error) {
+            console.info("File " + fileName + " was not uploaded (Time out, try increasing the maxUploadingTime value in the config)");
+            if(!config.application.tryToRecoverUnuploadedFiles) {
+                removeFileFromUploadList(fileName);
+                console.warn("File can't be upload next time as \"tryToRecoverUnuploadedFiles\" feature is disabled");
+            }
+        }
+    } else {
+        console.info("File " + fileName + " is not on the upload list, file can't be uploaded");
     }
+
 }
 
 /** Deleting file from hard drive */
 
-function deleteFile(fileName, fileId) {
+function deleteFile(fileName, fileId = NaN) {
     fs.unlinkSync(__dirname + "/downloads/" + fileName);
     console.info("File " + fileName + " deleted (Hard drive)");
     if(config.application.deletingFilesOnOneDriveAfterMoving) {
@@ -81,10 +120,40 @@ function deleteFile(fileName, fileId) {
 
 /** Deleting OneDrive Content (Optional) */
 
-async function deleteFileOnOneDrive(fileName, fileId) {
-    await oneDriveAPI.items.delete({
-        accessToken: oneDrive.accessToken,
-        itemId: fileId
-    })
-    console.info("File " + fileName + " deleted (OneDrive)");
+async function deleteFileOnOneDrive(fileName, fileId = NaN) {
+    if(fileId != NaN) {
+        await oneDriveAPI.items.delete({
+            accessToken: oneDrive.accessToken,
+            itemId: fileId
+        })
+        console.info("File " + fileName + " deleted (OneDrive)");
+    } else {
+        console.warn("File " + fileName + " can't be deleted (OneDrive) because no fileId is provided (probably because this file was not successfully uploaded last time)");
+    }
+}
+
+/** Recover function */
+
+const low = require('lowdb');
+const FileSync = require('lowdb/adapters/FileSync');
+ 
+const adapterA = new FileSync(__dirname + "/downloads/recovery.json");
+const adapter = low(adapterA);
+
+adapter.defaults = {"uploadList": []}
+
+function getUploadList() {
+    return adapter.get(`uploadList`);
+}
+
+function addFileToUploadList(fileName) {
+    let list = getUploadList();
+    list.push(fileName);
+    adapter.set(`uploadList`, list);
+}
+
+function removeFileFromUploadList(fileName) {
+    let list = getUploadList();
+    list.remove(fileName);
+    adapter.set(`uploadList`, list);
 }
